# lib/api-zod/src/generated/types/evento.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Denok Bat - API para asociación de jubilados, integración con Odoo
 * OpenAPI spec version: 0.1.0
 */

export interface Evento {
  id: number;
  nombre: string;
  /** @nullable */
  nombreEu?: string | null;
  /** @nullable */
  descripcion?: string | null;
  /** @nullable */
  descripcionEu?: string | null;
  fechaInicio: string;
  /** @nullable */
  fechaFin?: string | null;
  /** @nullable */
  lugar?: string | null;
  /** @nullable */
  imagen?: string | null;
  /** @nullable */
  plazasDisponibles?: number | null;
  inscrito: boolean;
}