# lib/api-zod/src/generated/types/noticiasList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Denok Bat - API para asociación de jubilados, integración con Odoo
 * OpenAPI spec version: 0.1.0
 */
import type { Noticia } from "./noticia";

export interface NoticiasList {
  items: Noticia[];
  total: number;
  page: number;
  limit: number;
}